Implemented discussed chages for PR #144 https://github.com/pg2455/covid_p2p_simulation/pull/144 Specifically, the changes are: - Fix for cached_property py<3.8 - conditional import & requirements.txt - Revert p_infection change ( removal of '* h.n_infectious_contacts')   - Thats for future commit/PR - Revised doc strings for is_dead, filter_*, expire, EmptyCity.initWorld - Revised filter_* -> filter_open, filter_queue_max - Use pathlib for accessing functional configs - Rename directory functional_configs -> test_configs - ExternalBuildCity renamed to EmptyCity and moved to base.py - remove filter_open!=[] in Human.run mobility block (its handled by Human.excursion)
