+++ requirements.txt
@@ -16,3 +16,4 @@ networkx
 tqdm
 seaborn
 pyyaml
+cached_property
