+++ src/covid19sim/utils.py
@@ -1529,10 +1529,18 @@ def calculate_average_infectiousness(human):
     return (cur_infectiousness + tomorrows_infectiousness) / 2
 
 def filter_open(locations):
-    """Given an iterable of locations, will return a list of those that are open for business"""
+    """Given an iterable of locations, returns a list of those that are open for business.
+    
+    Yields:
+        list
+    """
     return [loc for loc in locations if loc.is_open_for_business]
 
-def filter_open_and_short_queue(locations):
-    """Given an iterable of locations, will return a list of those that are open for business
-    and with queues that are not too long"""
-    return [loc for loc in locations if loc.is_open_for_business and len(loc.queue)<=MAX_STORE_QUEUE_LENGTH]
+def filter_queue_max(locations):
+    """Given an iterable of locations, will return a list of those
+    with queues that are not too long.
+    
+    Yields:
+        list
+    """
+    return [loc for loc in locations if len(loc.queue)<=MAX_STORE_QUEUE_LENGTH]
